requirements_build.txt
pyinstaller==6.3.0
opencv-python==4.8.1.78
numpy==1.26.2
pyautogui==0.9.54
moviepy==1.0.3
sounddevice==0.4.6
soundfile==0.12.1
pillow==10.1.0
setuptools>=65.5.1  # Required for PyInstaller
wheel>=0.38.0  # Required for building packages